Create new file for user routes
